old_FastApi/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.26
geoalchemy2==0.10.0
shapely==2.0.2
pydantic==1.10.2
python-dotenv==0.19.1